feat: improve readability
