+++ aiagents4pharma/talk2biomodels/pyproject.toml
@@ -40,6 +40,7 @@ dynamic = ["version"]
 packages = [
     "talk2biomodels",
     "talk2biomodels.models",
+    "talk2biomodels.api",
     "talk2biomodels.tools",
     "talk2biomodels.agents",
     "talk2biomodels.states"
