+++ pyproject.toml
@@ -68,3 +68,6 @@ include = ["aiagents4pharma*"]
 aiagents4pharma = [
             "configs/*",
             "configs/talk2biomodels/agents/t2b_agent/*"]
+
+[tool.pytest.ini_options]
+asyncio_default_fixture_loop_scope = "function"
